add options --skip-verify & --skip-mismatched-references
This introduces a verify step after the image was sucessfully written to ensure
the checksum is as expected, otherwise exit with error.
New option --skip-verify allows skipping this step (yielding behavior as before this change)

Also introduces the ability to skip reference images with
--skip-mismatched-references    if a reference image checksum
mismatches the full image checksum recorded in its
accompanying chunk index.
Without this option, such as mismatch will cause imidj to exit with error 64.