catch correct chunksum at different offset bug
Catch chunks where the chunksum and chunk # is the same in the pre-existing
image, but where the offset changed.

This can happen for example when a few bytes got added just before the
chunk.

In this case we cannot just skip it, either the chunk has to be relocated
(if its still intact) or re-written from another source at the new offset.

For now, always re-write from other source.